Delete old disable review apps files, rename new files, fix test names
