+++ packages/pipelines/test/commands/disable.test.ts
@@ -1,58 +1,223 @@
-import { expect, test } from '@oclif/test'
+import {expect, test} from '@oclif/test'
+// tslint:disable-next-line:no-duplicate-imports
 
-describe('reviewapps:disable', function () {
+describe('reviewapps:disable', () => {
   const pipeline = {
     id: '123-pipeline',
     name: 'my-pipeline'
   }
 
-  const app = {
-    id: '123-prod-app',
-    name: pipeline.name
-  }
+  describe('with repos api enabled', () => {
+    const feature = {
+      name: 'dashboard-repositories-api',
+      enabled: true
+    }
 
-  const testWithMocks = () => {
-    return test
-      .nock('https://kolkrabbi.heroku.com', api => {
-        api.patch(`/apps/${app.id}/github`).reply(200, {})
-      })
+    const repo = {
+      full_name: 'james/repo'
+    }
+
+    test
       .nock('https://api.heroku.com', api => {
-        api.get(`/apps/${app.name}`).reply(200, app)
+        api
+          .get(`/account/features/${feature.name}`)
+          .reply(200, feature)
+
+        api
+          .get(`/pipelines/${pipeline.name}`)
+          .reply(200, pipeline)
+        api
+          .get(`/pipelines/${pipeline.id}/repo`)
+          .reply(200, repo)
+        api
+          .post(`/pipelines/${pipeline.id}/review-app-config`)
+          .reply(200, {})
       })
-  }
+      .stderr()
+      .command(['reviewapps:disable', `--pipeline=${pipeline.name}`])
+      .it('succeeds with defaults', ctx => {
+        expect(ctx.stderr).to.include('Configuring pipeline')
+      })
 
-  testWithMocks()
-    .stderr()
-    .command(['reviewapps:disable', `--pipeline=${pipeline.name}`, `--app=${app.name}`])
-    .it('succeeds with defaults', ctx => {
-      expect(ctx.stderr).to.include('Configuring pipeline')
-    })
+    test
+      .nock('https://api.heroku.com', api => {
+        api
+          .get(`/account/features/${feature.name}`)
+          .reply(200, feature)
 
-  testWithMocks()
-    .stderr()
-    .stdout()
-    .command(['reviewapps:disable', `--pipeline=${pipeline.name}`, `--app=${app.name}`, '--autodeploy'])
-    .it('it disables autodeploys', ctx => {
-      expect(ctx.stdout).to.include('Disabling auto deployment')
-      expect(ctx.stderr).to.include('Configuring pipeline')
-    })
+        api
+          .get(`/pipelines/${pipeline.name}`)
+          .reply(200, pipeline)
+        api
+          .get(`/pipelines/${pipeline.id}/repo`)
+          .reply(200, repo)
+        api
+          .post(`/pipelines/${pipeline.id}/review-app-config`)
+          .reply(200, {})
+      })
+      .stdout()
+      .stderr()
+      .command(['reviewapps:disable', `--pipeline=${pipeline.name}`, '--autodeploy'])
+      .it('disables autodeploy', ctx => {
+        expect(ctx.stdout).to.include('Disabling auto deployment')
+        expect(ctx.stderr).to.include('Configuring pipeline')
+      })
 
-  testWithMocks()
-    .stderr()
-    .stdout()
-    .command(['reviewapps:disable', `--pipeline=${pipeline.name}`, `--app=${app.name}`, '--autodestroy'])
-    .it('it disables autodestroy', ctx => {
-      expect(ctx.stdout).to.include('Disabling auto destroy')
-      expect(ctx.stderr).to.include('Configuring pipeline')
-    })
+    test
+      .nock('https://api.heroku.com', api => {
+        api
+          .get(`/account/features/${feature.name}`)
+          .reply(200, feature)
 
-  testWithMocks()
-    .stderr()
-    .stdout()
-    .command(['reviewapps:disable', `--pipeline=${pipeline.name}`, `--app=${app.name}`, '--autodestroy', '--autodeploy'])
-    .it('it disables autodestroy and auto deploy', ctx => {
-      expect(ctx.stdout).to.include('Disabling auto destroy')
-      expect(ctx.stdout).to.include('Disabling auto deployment')
-      expect(ctx.stderr).to.include('Configuring pipeline')
-    })
+        api
+          .get(`/pipelines/${pipeline.name}`)
+          .reply(200, pipeline)
+        api
+          .get(`/pipelines/${pipeline.id}/repo`)
+          .reply(200, repo)
+        api
+          .post(`/pipelines/${pipeline.id}/review-app-config`)
+          .reply(200, {})
+      })
+      .stdout()
+      .stderr()
+      .command(['reviewapps:disable', `--pipeline=${pipeline.name}`, '--autodestroy'])
+      .it('disables autodestroy', ctx => {
+        expect(ctx.stdout).to.include('Disabling auto destroy')
+        expect(ctx.stderr).to.include('Configuring pipeline')
+      })
+
+    test
+      .nock('https://api.heroku.com', api => {
+        api
+          .get(`/account/features/${feature.name}`)
+          .reply(200, feature)
+
+        api
+          .get(`/pipelines/${pipeline.name}`)
+          .reply(200, pipeline)
+        api
+          .get(`/pipelines/${pipeline.id}/repo`)
+          .reply(200, repo)
+        api
+          .post(`/pipelines/${pipeline.id}/review-app-config`)
+          .reply(200, {})
+      })
+      .stdout()
+      .stderr()
+      .command(['reviewapps:disable', `--pipeline=${pipeline.name}`, '--autodeploy', '--autodestroy'])
+      .it('disables autodeploy and autodestroy', ctx => {
+        expect(ctx.stdout).to.include('Disabling auto deployment')
+        expect(ctx.stdout).to.include('Disabling auto destroy')
+        expect(ctx.stderr).to.include('Configuring pipeline')
+      })
+  })
+
+  describe('with repos api disabled', () => {
+    const feature = {
+      name: 'dashboard-repositories-api',
+      enabled: false
+    }
+
+    const repo = {
+      repository: {
+        name: 'james/repo'
+      }
+    }
+
+    test
+      .nock('https://api.heroku.com', api => {
+        api
+          .get(`/account/features/${feature.name}`)
+          .reply(200, feature)
+
+        api
+          .get(`/pipelines/${pipeline.name}`)
+          .reply(200, pipeline)
+        api
+          .get(`/pipelines/${pipeline.id}/repository`)
+          .reply(200, repo)
+        api
+          .post(`/pipelines/${pipeline.id}/review-app-config`)
+          .reply(200, {})
+      })
+      .stderr()
+      .command(['reviewapps:disable', `--pipeline=${pipeline.name}`])
+      .it('succeeds with defaults', ctx => {
+        expect(ctx.stderr).to.include('Configuring pipeline')
+      })
+
+    test
+      .nock('https://api.heroku.com', api => {
+        api
+          .get(`/account/features/${feature.name}`)
+          .reply(200, feature)
+
+        api
+          .get(`/pipelines/${pipeline.name}`)
+          .reply(200, pipeline)
+        api
+          .get(`/pipelines/${pipeline.id}/repository`)
+          .reply(200, repo)
+        api
+          .post(`/pipelines/${pipeline.id}/review-app-config`)
+          .reply(200, {})
+      })
+      .stdout()
+      .stderr()
+      .command(['reviewapps:disable', `--pipeline=${pipeline.name}`, '--autodeploy'])
+      .it('disables autodeploy', ctx => {
+        expect(ctx.stdout).to.include('Disabling auto deployment')
+        expect(ctx.stderr).to.include('Configuring pipeline')
+      })
+
+    test
+      .nock('https://api.heroku.com', api => {
+        api
+          .get(`/account/features/${feature.name}`)
+          .reply(200, feature)
+
+        api
+          .get(`/pipelines/${pipeline.name}`)
+          .reply(200, pipeline)
+        api
+          .get(`/pipelines/${pipeline.id}/repository`)
+          .reply(200, repo)
+        api
+          .post(`/pipelines/${pipeline.id}/review-app-config`)
+          .reply(200, {})
+      })
+      .stdout()
+      .stderr()
+      .command(['reviewapps:disable', `--pipeline=${pipeline.name}`, '--autodestroy'])
+      .it('disables autodestroy', ctx => {
+        expect(ctx.stdout).to.include('Disabling auto destroy')
+        expect(ctx.stderr).to.include('Configuring pipeline')
+      })
+
+    test
+      .nock('https://api.heroku.com', api => {
+        api
+          .get(`/account/features/${feature.name}`)
+          .reply(200, feature)
+
+        api
+          .get(`/pipelines/${pipeline.name}`)
+          .reply(200, pipeline)
+        api
+          .get(`/pipelines/${pipeline.id}/repository`)
+          .reply(200, repo)
+        api
+          .post(`/pipelines/${pipeline.id}/review-app-config`)
+          .reply(200, {})
+      })
+      .stdout()
+      .stderr()
+      .command(['reviewapps:disable', `--pipeline=${pipeline.name}`, '--autodeploy', '--autodestroy'])
+      .it('disables autodeploy and autodestroy', ctx => {
+        expect(ctx.stdout).to.include('Disabling auto deployment')
+        expect(ctx.stdout).to.include('Disabling auto destroy')
+        expect(ctx.stderr).to.include('Configuring pipeline')
+      })
+  })
 })
